Make width and height of the OpenLayers directive a dynamic scope attribute.
The current OpenLayers directive does not allow updates on the width and height attributes (data binding). As this is sometimes necessary, putting these attributes into the scope should be useful.

The updateSize() function of the OpenLayers Map class notifies the Viewport such that it can adapt.

Static values need to be wrapped in single quotation marks. An example:

<openlayers width="'100%'" height="customHeight"></openlayers>