Added index to layers and optionally to properties
If the user adds an index property to the ol-layer-properties,
the code now tries to maintain that layer number. It only enforces
the index when
1) the layer is added
2) the layer properties (or index) are modified
The changes require all code to use olHelpers code to access the
ol.Map.getLayers() layer collection.

The use of an index by the user is optional. I update the underlying
array as follows:

JavaScript:
$scope.layers = [layer1, layer2]
$scope.updateLayerOrder = function() {
    $scope.layers.map(function(layer, idx){ layer.index = idx; });
}

HTML
<openlayers custom-layers="true">
<ol-layer ng-repeat="layer in layers track by layer.name" ol-layer-properties="layer"></ol-layer>
</openlayers>

An alternative implementation that I considered: watch the ol-layer element index compared to
its siblings, and use that index. This requires less change from the user, but I couldn't find
a way to integrate it in the code.