Support keepOneOverlayVisible feature in markers
Feature to show a marker's overlay until the user enters the mouse in other marker, when the overlay is changed to display the new marker's info. With this new behaviour an overlay is always visible unless the user clicks on a free marker zone of the map, which forces the visible overlay to close itself.