feat(olMarkers): Added support for marker dragging
Adding a boolean property 'draggable' to the marker can make it movable or not by the user.
This change actually also fixed a bug about changing dynamically the show label options which was not correctly handled due to a bad management of event listeners.
I updated the 063-markers-properties-example to show all of this.